setup batch predict feature and related documentation

+++ scripts/batch_predict/batch_predict.sh
@@ -1,7 +1,21 @@
 #!/bin/bash
 
-docker-compose up -d
 
-curl --header "Content-Type: application/json" --request POST --data-binary @../../ml/input/api/payload.json http://localhost:8080/invocations > output.json
+REQUIRED_FILE1="../../ml/input/api/payload.json"
+REQUIRED_FILE2="../../ml/output/credit-card-fraud/model.joblib" 
 
-docker-compose down
+if [ -f $REQUIRED_FILE1 ] && [ -f $REQUIRED_FILE2 ]; then
+    docker-compose up -d
+
+    while [ ! -f response.json ]
+    do
+    sleep 0.2
+    done
+
+    ls -l response.json
+
+    docker-compose down
+else
+   echo "Files $REQUIRED_FILE1 and $REQUIRED_FILE2 does not exist."
+fi
+
